+++ docu/changelog.md
@@ -28,8 +28,8 @@ Navigate down to logs area and invoke the VIP filter.<br>
 <br></li>
 <li>
 Change: [Issue <a href="https://github.com/2Abendsegler/GClh/issues/264">#264</a>] Show button "Show log counter" only if feature is possible:<br>
-Do not show button "Show log counter" if the logs are filtered or searched.
-</li>
+Do not show button "Show log counter" if the logs are filtered or searched.<br>
+<br></li>
 <li>
 Fix: [Issue <a href="https://github.com/2Abendsegler/GClh/issues/264">#264</a>] VIP list not rendered on a specific cache (Lesopark Dubina). Profile link in listing without any text.<br>
 </li>