+++ docu/changelog.md
@@ -27,7 +27,7 @@ Navigate down to logs area and invoke the VIP filter.<br>
 <img src="../images/v0.7_Screen02.jpg" alt="v0.7_Screen02.jpg"><br>
 <br></li>
 <li>
-Change: [Issue <a href="https://github.com/2Abendsegler/GClh/issues/264">#264</a>] Show button "Show log counter" only if feature is possible:<br>
+Change: [Issue <a href="https://github.com/2Abendsegler/GClh/issues/277">#277</a>] Show button "Show log counter" only if feature is possible:<br>
 Do not show button "Show log counter" if the logs are filtered or searched.<br>
 </li>
 <li>