+++ docu/changelog.md
@@ -35,7 +35,7 @@ Fix: [Issue <a href="https://github.com/2Abendsegler/GClh/issues/264">#264</a>] 
 </li>
 <br>
 </ul>
-(??.04.2017)<br> 
+(07.04.2017)<br> 
 <br>
 
 ---